64-bit limits (max_vm_size, etc) are used in compile packet

+++ compile_packet_1.c
@@ -82,9 +82,9 @@ compile_request_packet_read(
   pout->ts1_us = cvt_bin_to_host_32(pin->ts1_us);
   FAIL_IF(pout->ts1_us < 0 || pout->ts1_us > USEC_MAX);
 
-  pout->max_vm_size = (size_t) cvt_bin_to_host_64(pin->max_vm_size);
-  pout->max_stack_size = (size_t) cvt_bin_to_host_64(pin->max_stack_size);
-  pout->max_file_size = (size_t) cvt_bin_to_host_64(pin->max_file_size);
+  pout->max_vm_size = cvt_bin_to_host_64(pin->max_vm_size);
+  pout->max_stack_size = cvt_bin_to_host_64(pin->max_stack_size);
+  pout->max_file_size = cvt_bin_to_host_64(pin->max_file_size);
 
   pout->use_uuid = cvt_bin_to_host_32(pin->use_uuid);
   pout->uuid.v[0] = cvt_bin_to_host_32(pin->uuid.v[0]);
